Move all public *Tools.h in src.
This by definition remove all the tool functions from the public API.
