+++ include/meson.build
@@ -17,15 +17,3 @@ headers = [
 ]
 
 install_headers(headers, subdir:'kiwix')
-
-install_headers(
-  'tools/base64.h',
-  'tools/networkTools.h',
-  'tools/otherTools.h',
-  'tools/pathTools.h',
-  'tools/regexTools.h',
-  'tools/stringTools.h',
-  'tools/archiveTools.h',
-  subdir:'kiwix/tools'
-)
-
