+++ src/book.cpp
@@ -24,6 +24,8 @@
 #include "tools/regexTools.h"
 #include "tools/networkTools.h"
 #include "tools/otherTools.h"
+#include "tools/stringTools.h"
+#include "tools/pathTools.h"
 
 #include <pugixml.hpp>
