+++ src/tools/pathTools.h
