+++ src/tools/regexTools.h
