+++ src/tools/stringTools.h
