+++ test/server.cpp
@@ -5,7 +5,7 @@
 #include "../include/manager.h"
 #include "../include/server.h"
 #include "../include/name_mapper.h"
-
+#include "../src/tools/pathTools.h"
 
 using TestContextImpl = std::vector<std::pair<std::string, std::string> >;
 struct TestContext : TestContextImpl {
